Avoid copies during sorting.
This provides an ~4% improvement in sort performance, but having benchmarked it, sorting isn't really worth optimizing, as it's a small fraction of total time spent.

This also eliminates some code by relying on the permutation crate, and is no more complex, so is worth landing despite the negligible performance win.